Moves acceptance tests to proper location.
