+++ finalApp/acceptance_tests/test_editCourse.py
