+++ templates/account_list.html
@@ -12,7 +12,12 @@
     <form action="" method="post">
         {% csrf_token %}
         <!-- insert selection list box here -->
-
+        <select size="4" id="accounts" name="accounts" onclick="window.location.href='/account_list/';">
+                {% for i in accounts  %}
+                 <option value="{{ i.id }}">{{ i.title }}</option>
+                {% endfor %}
+            </select><br><br>
+        </select>
     </form>
     <form align="right" method="post">
         <label class="logoutLblPos">
